% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_grouping.R
\name{add_grouping_bins}
\alias{add_grouping_bins}
\title{Add grouping based on time bins}
\usage{
add_grouping_bins(
  data,
  label = "group",
  dictionary = list(time = "TIME"),
  bins = c(0, 24, 48, 72, Inf)
)
}
\arguments{
\item{data}{NONMEM-style data.frame, or path to CSV file with NONMEM data}

\item{label}{column name of grouper column in dataset}

\item{dictionary}{data dictionary. Optional, a named character vector that specifies the column
names to be used from the dataset.}

\item{bins}{vector of bin separators}
}
\description{
Add grouping based on time bins
}
