% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_metrics.R
\name{bootstrap_metrics}
\alias{bootstrap_metrics}
\alias{summarise_bootstrap_metrics}
\alias{summarize_bootstrap_metrics}
\title{Calculate metrics across bootstrapped folds}
\usage{
bootstrap_metrics(.data, ..., .by = NULL, .seed = 10, .n_boots = 1000)

summarise_bootstrap_metrics(.data, .by = NULL, .conf_level = 0.95)

summarize_bootstrap_metrics(.data, .by = NULL, .conf_level = 0.95)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g. a tibble).}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Name-value pairs of
summary functions to compute across bootstrapped folds of the data. The
name will be the name of the variable in the result.}

\item{.by}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Optionally, variables to
group by. Each group is resampled independently.}

\item{.seed}{Single value for the random seed, used for reproducible random
sampling.}

\item{.n_boots}{Number of bootstrapped samples to create (per group).}

\item{.conf_level}{The confidence level to use for the confidence interval.
Must be strictly between 0 and 1. Defaults to a 95 percent confidence
interval.}
}
\value{
For \code{bootstrap_metrics()}: A \code{\link[tibble:tibble]{tibble::tibble()}} whose columns are a
combination of the summary expressions and grouping keys that you provide,
plus a \code{boot} column indicating each bootstrap fold.

For \code{summarise_bootstrap_metrics()}: A \code{\link[tibble:tibble]{tibble::tibble()}} with, for each
bootstrapped metric, three columns suffixed \verb{_mid}, \verb{_lower}, and \verb{_upper},
giving the mean and the lower and upper quantiles implied by \code{.conf_level},
respectively.
}
\description{
Calculate metrics across bootstrapped folds of the data; optionally, according
to one or more grouped variables. Bootstrap size is equal to the size of the
data or each grouped variable, and rows are sampled with replacement.
}
\section{Useful metrics}{

\itemize{
\item (Normalized) Root mean squared error: \code{\link[=rmse]{rmse()}}, \code{\link[=nrmse]{nrmse()}}
\item Mean (absolute) percent error: \code{\link[=mpe]{mpe()}}, \code{\link[=mape]{mape()}}
\item Accuracy: \code{\link[=accuracy]{accuracy()}}
}
}

\examples{
set.seed(99)
df <- data.frame(
  observationid = rep(1:1000, 4),
  model = rep(c(rep("A", 1000), rep("B", 1000)), 2),
  patient_type = "general",
  prediction_type = c(rep("a priori", 2000), rep("a posteriori", 2000)),
  res = c(
    rnorm(1000, 2, 3),
    rnorm(1000, 0.1, 1),
    rnorm(1000, 1, 3),
    rnorm(1000, 0, 0.5)
  ),
  tdm = rnorm(4000, 10, 5)
)

boots <- bootstrap_metrics(
  df,
  rmse = rmse(tdm, tdm - res),
  nrmse = nrmse(tdm, tdm - res),
  accuracy = accuracy(tdm, tdm - res, 2.5, 0.2),
  .by = c(model, patient_type, prediction_type),
  .n_boots = 100
)

summarise_bootstrap_metrics(
  boots,
  .by = c(model, patient_type, prediction_type)
)
}
