% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_shrinkage.R
\name{calc_eta}
\alias{calc_eta}
\title{Calculate the "eta"-value for a parameters, assuming an exponential
shape for IIV, i.e. \code{PAR = TV_PAR * EXP(ETA(n))}, and
additive shape for IOV (kappa's), i.e. \code{PAR = TV_PAR + ETA(n)} or when
final parameter estimate is 0.}
\usage{
calc_eta(ind, par_name, parameters)
}
\arguments{
\item{ind}{individual parameter estimate}

\item{par_name}{name of parameter in model}

\item{parameters}{list of parameter values, should include entry for
\code{par_name}}
}
\description{
Calculate the "eta"-value for a parameters, assuming an exponential
shape for IIV, i.e. \code{PAR = TV_PAR * EXP(ETA(n))}, and
additive shape for IOV (kappa's), i.e. \code{PAR = TV_PAR + ETA(n)} or when
final parameter estimate is 0.
}
