% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bayesian_impact.R
\name{calculate_bayesian_impact}
\alias{calculate_bayesian_impact}
\title{This function calculates the impact of using Bayesian updating, compared
to using population estimates}
\usage{
calculate_bayesian_impact(stats_summ)
}
\arguments{
\item{stats_summ}{stats summary, a data.frame created in \code{run_eval()}}
}
\value{
a tibble with bayesian impact values, based on RMSE and MAPE
}
\description{
Bayesian impact (BI), as defined here, is the gain in accuracy that is
obtained when using Bayesian updating, compared to using a priori
(population) parameter estimates. The BI is computed based on accuracy
measures (RMSE and MAPE), but not on bias estimates (MPE).
}
