% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stats.R
\name{calculate_stats}
\alias{calculate_stats}
\title{Calculate basic statistics, like RMSE, MPE, MAPE for forecasted data}
\usage{
calculate_stats(.res, rounding = 3, acc_error_abs = NULL, acc_error_rel = NULL)
}
\arguments{
\item{.res}{output object (\code{mipdeval_results}) from \code{run_eval()}, or
\code{data.frame} with raw results.}

\item{rounding}{number of decimals to round to.}

\item{acc_error_abs, acc_error_rel}{For calculating \code{\link[=accuracy]{accuracy()}}: Positive number
providing an absolute or relative error margin. The cutoff is exclusive of
the error margin. When \code{NULL} (the default), accuracy will not be
calculated and will return \code{NA} instead.}
}
\value{
tibble
}
\description{
Calculate basic statistics, like RMSE, MPE, MAPE for forecasted data
}
