% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_data.R
\name{check_input_data}
\alias{check_input_data}
\title{Do some checks and minor manipulations on input dataset}
\usage{
check_input_data(data, dictionary)
}
\arguments{
\item{data}{NONMEM-style data.frame, or path to CSV file with NONMEM data}

\item{dictionary}{data dictionary. Optional, a named character vector that specifies the column
names to be used from the dataset.}
}
\value{
a data.frame
}
\description{
E.g. confirm that all required columns are included, and column names are
all lower-case.
}
