% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_model.R
\name{check_installed_model_library}
\alias{check_installed_model_library}
\title{Check if PKPDsim model library is installed}
\usage{
check_installed_model_library(model, call = rlang::caller_env())
}
\arguments{
\item{model}{Character. Name of an installed PKPDsim model library.}

\item{call}{The execution environment of a currently running
function, e.g. \code{call = caller_env()}. The corresponding function
call is retrieved and mentioned in error messages as the source
of the error.

You only need to supply \code{call} when throwing a condition from a
helper function which wouldn't be relevant to mention in the
message.

Can also be \code{NULL} or a \link[rlang:topic-defuse]{defused function call} to
respectively not display any call or hard-code a code to display.

For more information about error calls, see \ifelse{html}{\link[rlang:topic-error-call]{Including function calls in error messages}}{\link[rlang:topic-error-call]{Including function calls in error messages}}.}
}
\value{
This function is called for its side effects and returns \code{NULL} if
the PKPDsim model library is installed or returns an error otherwise.
}
\description{
Check if PKPDsim model library is installed
}
