% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_psn_results.R
\name{compare-psn-results}
\alias{compare-psn-results}
\alias{compare_psn_proseval_results}
\alias{reldiff_psn_proseval_results}
\alias{compare_psn_execute_results}
\alias{reldiff_psn_execute_results}
\title{Compare mipdeval results with PsN}
\usage{
compare_psn_proseval_results(mipdeval, psn, tol = 0.1)

reldiff_psn_proseval_results(mipdeval, psn)

compare_psn_execute_results(mipdeval, psn, tol = 0.1)

reldiff_psn_execute_results(mipdeval, psn)
}
\arguments{
\item{mipdeval}{The \code{run_eval()} object.}

\item{psn}{A PsN execute or proseval data set.}

\item{tol}{Optional. Tolerance.}
}
\value{
A data/frame.
}
\description{
The \verb{compare_psn_*_results()} and \verb{reldiff_psn_*_results()} family of
functions are designed to compare mipdeval results with PsN:
\itemize{
\item \code{compare_psn_proseval_results()} and \code{compare_psn_execute_results()} summarize
the relative difference in predictions between mipdeval and the PsN \code{execute}
and \code{proseval} tools.
\item \code{reldiff_psn_proseval_results()} and \code{reldiff_psn_execute_results} calculate
the relative difference for each prediction between mipdeval and the PsN
\code{execute} and \code{proseval} tools.
}
}
