% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eval.R
\name{fit_options}
\alias{fit_options}
\title{Options for controlling MAP Bayesian fit}
\usage{
fit_options(..., reltol = 1e-05)
}
\arguments{
\item{...}{These dots are reserved for future extensibility and must be empty.}

\item{reltol}{Relative convergence tolerance. \code{reltol = 1e-04} will perform a
slightly coarser but more stable fit, which can be useful in some case.}
}
\value{
A list.
}
\description{
Options for controlling MAP Bayesian fit
}
