% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_shrinkage.R
\name{get_omega_for_parameters}
\alias{get_omega_for_parameters}
\title{Return all diagonal om^2 elements for each non-fixed parameter, as a list}
\usage{
get_omega_for_parameters(mod_obj)
}
\arguments{
\item{mod_obj}{list object with model information}
}
\value{
list
}
\description{
Return all diagonal om^2 elements for each non-fixed parameter, as a list
}
