% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_required_covariates}
\alias{get_required_covariates}
\title{Get required covariates from a PKPDsim model object}
\usage{
get_required_covariates(model)
}
\arguments{
\item{model}{PKPDsim model object}
}
\value{
A character vector of covariate names.
}
\description{
Get required covariates from a PKPDsim model object
}
