% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_grouping_column.R
\name{group_by_dose}
\alias{group_by_dose}
\title{Will create a separate group for each dose intervals that contains at least
one sample}
\usage{
group_by_dose(
  data,
  dictionary = list(ID = "ID", EVID = "EVID", TIME = "TIME"),
  ...
)
}
\arguments{
\item{data}{NONMEM-style data.frame, or path to CSV file with NONMEM data}

\item{dictionary}{data dictionary. Optional, a named character vector that
specifies the column names to be used from the dataset.}

\item{...}{Additional arguments passed onto \code{fun}. See the respective
grouping functions for more details. E.g. for \code{group_by_time}, need a
\code{bins} argument.}
}
\value{
a numeric vector of the same length as the number of rows in \code{data}.
}
\description{
Will create a separate group for each dose intervals that contains at least
one sample
}
\examples{
group_by_dose(nm_busulfan)

}
