% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_grouping_column.R
\name{group_by_time}
\alias{group_by_time}
\title{Group data by time using bin separators}
\usage{
group_by_time(
  data,
  bins = c(0, 24, 48, 72, 96, Inf),
  dictionary = list(TIME = "TIME"),
  ...
)
}
\arguments{
\item{data}{NONMEM-style data.frame, or path to CSV file with NONMEM data}

\item{bins}{vector of bin separators. Suggestion to keep the last bin
separator \code{Inf}, to ensure all points are included in a group.}

\item{...}{passed onto the chosen grouper function, see the respective
grouping functions for more details. E.g. for \code{group_by_time}, need a \code{bins}
argument.}
}
\value{
a vector of the same length as the number of rows in \code{.data}
}
\description{
Group data by time using bin separators
}
