% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eval_core.R
\name{handle_covariate_censoring}
\alias{handle_covariate_censoring}
\title{Handle covariate censoring}
\usage{
handle_covariate_censoring(covariates, t, censor = TRUE)
}
\arguments{
\item{covariates}{covariates data for single subject}

\item{t}{timepoint cutoff at which covariate are to be censored
(if \code{censor=FALSE})}

\item{censor}{censor covariate data?}
}
\value{
list with similar shape as \code{covariates} object, just
with potentially censored future data.
}
\description{
This removes new covariate information after a certain time cutoff
}
