% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_timevarying}
\alias{is_timevarying}
\title{For requested columns in a dataset, check if values vary or not across
rows}
\usage{
is_timevarying(.data, .cols)
}
\arguments{
\item{.data}{data.frame}

\item{.cols}{vector of column names}
}
\value{
A data.frame
}
\description{
For requested columns in a dataset, check if values vary or not across
rows
}
