% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mape}
\alias{mape}
\title{Mean absolute prediction error}
\usage{
mape(obs, pred)
}
\arguments{
\item{obs}{observations vector}

\item{pred}{predictions vector}
}
\value{
A numeric vector
}
\description{
Mean absolute prediction error
}
