% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipdeval-package.R
\docType{package}
\name{mipdeval-package}
\alias{mipdeval}
\alias{mipdeval-package}
\title{mipdeval: Run evaluation of iterative predictive performance based on historical datasets}
\description{
The package provides functionality to run through a NONMEM-style dataset iteratively, performing MAP Bayesian updating and forecasting at each step. This is very much like the `proseval` tool in PsN, however this package provides more options, specifically for use in MIPD, such as flattening priors, time-based weighting, sample grouping, etc.
}
\author{
\strong{Maintainer}: Ron Keizer \email{ron@insight-rx.com}

Authors:
\itemize{
  \item Michael McCarthy \email{michael.mccarthy@insight-rx.com}
}

Other contributors:
\itemize{
  \item InsightRX [copyright holder, funder]
}

}
\keyword{internal}
