% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mpe}
\alias{mpe}
\title{Mean prediction error}
\usage{
mpe(obs, pred)
}
\arguments{
\item{obs}{observations vector}

\item{pred}{predictions vector}
}
\value{
A numeric vector
}
\description{
Mean prediction error
}
