% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nm_busulfan}
\alias{nm_busulfan}
\title{Busulfan data}
\format{
A NONMEM-style data frame with 240 rows and
14 columns:

\describe{
\item{\code{ID}}{Patient ID.}
\item{\code{TIME}}{The recorded times of dosing events and/or observations.}
\item{\code{CMT}}{Dose compartment.}
\item{\code{DV}}{The dependent variable or observed concentration.}
\item{\code{AMT}}{Dose administered for dosing record.}
\item{\code{EVID}}{Event ID (0 = observation, 1 = dose, 2 = other, 3 = reset, 4 = reset + dose).}
\item{\code{MDV}}{Missing DV (0 = not missing, 1 = missing).}
\item{\code{RATE}}{Rate of drug infusion.}
\item{\code{AGE}}{Patient age at TIME of measurement.}
\item{\code{WT}}{Patient weight at TIME of measurement.}
\item{\code{HT}}{Patient height at TIME of measurement.}
\item{\code{SEX}}{Patient sex (0 = female, 1 = male).}
\item{\code{REGI}}{Regimen type.}
}
}
\usage{
nm_busulfan
}
\description{
Busulfan data simulated with the \code{"pkbusulfanshukla"} model.
}
\references{
Shukla et al. Front Pharmacol. 2020.
\url{https://doi.org/10.3389/fphar.2020.00888}
}
\keyword{datasets}
