% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nm_vanco}
\alias{nm_vanco}
\title{Vancomycin data}
\format{
A NONMEM-style data frame with 80 rows and
11 columns:

\describe{
\item{\code{ID}}{Patient ID.}
\item{\code{TIME}}{The recorded times of dosing events and/or observations.}
\item{\code{CMT}}{Dose compartment.}
\item{\code{DV}}{The dependent variable or observed concentration.}
\item{\code{AMT}}{Dose administered for dosing record.}
\item{\code{EVID}}{Event ID (0 = observation, 1 = dose, 2 = other, 3 = reset, 4 = reset + dose).}
\item{\code{MDV}}{Missing DV (0 = not missing, 1 = missing).}
\item{\code{RATE}}{Rate of drug infusion.}
\item{\code{WT}}{Patient weight at TIME of measurement.}
\item{\code{CRCL}}{Patient creatinine clearance at TIME of measurement.}
\item{\code{CL_HEMO}}{Additional clearance attributable to hemodialysis.}
}
}
\usage{
nm_vanco
}
\description{
Vancomycin data simulated with the \code{"pkvancothomson"} model.
}
\references{
Thomson et al. J Antimicrob Chemotherap. 2009.
\url{https://doi.org/10.1093/jac/dkp085}
}
\keyword{datasets}
