% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{nrmse}
\alias{nrmse}
\title{Normalized root-mean-squared error}
\usage{
nrmse(obs, pred)
}
\arguments{
\item{obs}{observations vector}

\item{pred}{predictions vector}
}
\value{
A numeric vector
}
\description{
Normalized root-mean-squared error
}
