% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_model.R
\name{parse_model}
\alias{parse_model}
\alias{parse_model.character}
\alias{parse_model.PKPDsim}
\title{Parse PKPDsim model information}
\usage{
parse_model(model, ...)

\method{parse_model}{character}(
  model,
  ...,
  parameters = NULL,
  ruv = NULL,
  omega = NULL,
  fixed = NULL,
  iov = NULL
)

\method{parse_model}{PKPDsim}(model, ..., parameters, ruv, omega, fixed = NULL, iov = NULL)
}
\arguments{
\item{model}{either a PKPDsim model object, or a string pointing to a
PKPDsim-generated model library, e.g. \code{pkvancothomson}}

\item{...}{Arguments passed to methods.}

\item{parameters}{list of parameters}

\item{ruv}{residual error variability magnitude, specified as list.}

\item{omega}{between subject variability, supplied as vector specifiying the
lower triangle of the covariance matrix of random effects}

\item{fixed}{fix a specific parameters, supplied as vector of strings}

\item{iov}{a list specifying the required metadata for implementation of IOV,
specifically the CV of the IOV for each parameter and a vector of bin
separators. For example,
\code{list(cv = list(CL = 0.1, V = 0.2), bins = c(0, 24, 48, 9999))}}
}
\value{
A named list.
}
\description{
Parse model information from PKPDsim model object or an installed PKPDsim
model library.
}
