% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_model_info.R
\name{parse_model_info}
\alias{parse_model_info}
\title{Parse model info based on user-supplied info and potentially PKPDsim library}
\usage{
parse_model_info(model, parameters, omega, ruv, iov_bins)
}
\arguments{
\item{model}{either a PKPDsim model object, or a string pointing to a
PKPDsim-generated model library, e.g. \code{pkvancothomson}}

\item{parameters}{list of parameters}

\item{omega}{between subject variability, supplied as vector specifiying the
lower triangle of the covariance matrix of random effects}

\item{ruv}{residual error variability magnitude, specified as list.}

\item{iov_bins}{bins for inter-occasion variability. Passed unchanged to 
PKPDsim.}
}
\description{
Parse model info based on user-supplied info and potentially PKPDsim library
}
