% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mipdeval_results}
\alias{plot.mipdeval_results}
\title{Plot method for a \code{run_eval()} object}
\usage{
\method{plot}{mipdeval_results}(x, type = "vpc", ...)
}
\arguments{
\item{x}{An object.}

\item{type}{Character string, indicating the type of plot. Options are \code{"vpc"}.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A plot.
}
\description{
One plot \code{type} is currently available:
\itemize{
\item \code{type = "vpc"} plots visual predictive checks (VPC) with \code{\link[vpc:vpc]{vpc::vpc()}}.
}
}
