% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.mipdeval_results}
\alias{print.mipdeval_results}
\title{Print results from run_eval()}
\usage{
\method{print}{mipdeval_results}(x, ...)
}
\arguments{
\item{x}{output object from \code{run_eval()}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print results from run_eval()
}
