% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.mipdeval_results_bayesian_impact}
\alias{print.mipdeval_results_bayesian_impact}
\title{Print Bayesian impact results from run_eval()}
\usage{
\method{print}{mipdeval_results_bayesian_impact}(x, ...)
}
\arguments{
\item{x}{bayesian impact object in output from \code{run_eval()}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print Bayesian impact results from run_eval()
}
