% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.mipdeval_results_shrinkage}
\alias{print.mipdeval_results_shrinkage}
\title{Print shrinkage results from run_eval()}
\usage{
\method{print}{mipdeval_results_shrinkage}(x, ...)
}
\arguments{
\item{x}{shrinkage object in output from \code{run_eval()}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print shrinkage results from run_eval()
}
