% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.mipdeval_results_stats_summ}
\alias{print.mipdeval_results_stats_summ}
\title{Print predictive performance statistics from run_eval()}
\usage{
\method{print}{mipdeval_results_stats_summ}(x, ...)
}
\arguments{
\item{x}{stats summary object in output from \code{run_eval()}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print predictive performance statistics from run_eval()
}
