% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input_data.R
\name{read_input_data}
\alias{read_input_data}
\title{Read input data}
\usage{
read_input_data(data, verbose = TRUE)
}
\arguments{
\item{data}{NONMEM-style data.frame, or path to CSV file with NONMEM data}

\item{verbose}{show more output}
}
\value{
data.frame
}
\description{
Read input data
}
