% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{install_default_literature_model}
\alias{new_ode_model}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{PKPDsim}{\code{\link[PKPDsim]{install_default_literature_model}}, \code{\link[PKPDsim]{new_ode_model}}}
}}

