% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rmse}
\alias{rmse}
\title{Root-mean-squared error}
\usage{
rmse(obs, pred)
}
\arguments{
\item{obs}{observations vector}

\item{pred}{predictions vector}
}
\value{
A numeric vector
}
\description{
Root-mean-squared error
}
