% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eval_core.R
\name{run_eval_core}
\alias{run_eval_core}
\title{Core iterative simulation and MAP estimation function that loops over
an individual's dataset}
\usage{
run_eval_core(
  mod_obj,
  data,
  weights = NULL,
  weight_prior = 1,
  censor_covariates = TRUE,
  incremental = FALSE,
  progress_function = function() {
 },
  .fit_options = NULL
)
}
\arguments{
\item{mod_obj}{list object with model information}

\item{data}{NONMEM-style data.frame, or path to CSV file with NONMEM data}

\item{weights}{optional sample downweighting scheme based on how long ago
observations were taken. Either a string naming the scheme (e.g.
\code{"weight_gradient_exponential"}), or a named list with a \code{scheme} element
plus any scheme-specific parameters (e.g.
\code{list(scheme = "weight_gradient_exponential", t12_decay = 72)}). See
\code{\link[=calculate_fit_weights]{calculate_fit_weights()}} for all available schemes and their parameters.
Default is \code{NULL} (no downweighting; all included samples are weighted
equally).}

\item{weight_prior}{weighting of priors in relationship to observed data, 
default = 1}

\item{censor_covariates}{with the \code{proseval} tool in PsN, there is “data
leakage” (of future covariates data): since the NONMEM dataset in each step
contains the covariates for the future, this is technically data leakage,
and could result in an over-optimistic estimate of predictive performance.
In \code{mipdeval}, covariate censoring of future covariate data is switched on
by default (so no data leakage), but it can be switched off if we want to
match the behavior of \code{PsN::proseval} exactly.}

\item{incremental}{should MAP Bayesian do incremental fits in the iterative
loop? I.e. in this case it would use the first iterations MAP Bayesian
estimates as input for the second iteration, and so forth. The uncertainty
around the MAP estimates would be used as the new \code{omega} matrix. This
approach has been called "model predictive control (MPC)"
(www.page-meeting.org/?abstract=9076) and may be more predictive than
"regular" MAP in some scenarios. Default is \code{FALSE}.}

\item{progress_function}{function to increment progress bar}

\item{.fit_options}{Options for controlling MAP Bayesian fit. This must be
the result from a call to \code{\link[=fit_options]{fit_options()}}.}
}
\value{
a \code{data.frame} with individual predictions
}
\description{
Core iterative simulation and MAP estimation function that loops over
an individual's dataset
}
\keyword{internal}
