% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vpc.R
\name{run_vpc}
\alias{run_vpc}
\title{Run iterative predictive analysis, looping over each individual's data
to create observation and simulation datasets for creating visual predictive
checks (VPCs).}
\usage{
run_vpc(
  model,
  data,
  ids = NULL,
  n_samples = 250,
  parameters = NULL,
  fixed = NULL,
  omega = NULL,
  iov = NULL,
  ruv = NULL,
  dictionary = list(),
  group = NULL,
  progress = TRUE,
  threads = 1,
  verbose = TRUE
)
}
\arguments{
\item{model}{either a PKPDsim model object, or a string pointing to a
PKPDsim-generated model library, e.g. \code{pkvancothomson}}

\item{data}{NONMEM-style data.frame, or path to CSV file with NONMEM data}

\item{ids}{optional, vector of subject IDs to run analysis on (by default
runs analysis on all subjects in dataset)}

\item{n_samples}{number of iterations to perform for VPC simulations.}

\item{parameters}{list of parameters}

\item{fixed}{fix a specific parameters, supplied as vector of strings}

\item{omega}{between subject variability, supplied as vector specifiying the
lower triangle of the covariance matrix of random effects}

\item{iov}{a list specifying the required metadata for implementation of IOV,
specifically the coefficient of variation (CV \%) of the IOV for each
parameter and a vector of bin separators. For example,
\code{list(cv = list(CL = 0.1, V = 0.2), bins = c(0, 24, 48, 9999))}}

\item{ruv}{residual error variability magnitude, specified as list.}

\item{dictionary}{data dictionary. Optional, a named character vector that
specifies the column names to be used from the dataset.}

\item{group}{name of column in \code{data} that groups observations together in
iterative flow. By default each observation will be its own "group", but
this can be used to group peaks and troughs together, or to group
observations on the same day together. Grouping will be done prior to
running the analysis, so cannot be changed afterwards.}

\item{progress}{should a progress bar be shown? Default is \code{TRUE}, but when
debugging the package it is useful to have it off, since progress bar
handlers obscure R output.}

\item{threads}{number of threads to divide computations on. Default is 1,
i.e. no parallel execution}

\item{verbose}{show more output}
}
\value{
A named list of data frames with \code{obs} and \code{sim} data.
}
\description{
Run iterative predictive analysis, looping over each individual's data
to create observation and simulation datasets for creating visual predictive
checks (VPCs).
}
\examples{
\dontrun{
library(mipdeval)
library(vpc)
vpc_vanc <- run_vpc(
  model = "pkvancothomson",
  data = nm_vanco,
  n_samples = 100,
  progress = FALSE
)
vpc(
  sim = vpc_vanc$sim,
  obs = vpc_vanc$obs
)
}

}
