% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ss}
\alias{ss}
\title{Weighted sum-of-squares of residuals}
\usage{
ss(obs, pred, w = NULL)
}
\arguments{
\item{obs}{observations vector}

\item{pred}{predictions vector}

\item{w}{weights}
}
\description{
Weighted sum-of-squares of residuals
}
