% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stats.R
\name{stats_summ_options}
\alias{stats_summ_options}
\title{Options for summary statistics}
\usage{
stats_summ_options(
  ...,
  rounding = 3,
  acc_error_abs = NULL,
  acc_error_rel = NULL,
  bootstrap = bootstrap_options()
)
}
\arguments{
\item{...}{These dots are reserved for future extensibility and must be empty.}

\item{rounding}{number of decimals to round to.}

\item{acc_error_abs, acc_error_rel}{For calculating \code{\link[=accuracy]{accuracy()}}: Positive number
providing an absolute or relative error margin. The cutoff is exclusive of
the error margin. When \code{NULL} (the default), accuracy will not be
calculated and will return \code{NA} instead.}

\item{bootstrap}{Options for bootstrapping confidence intervals around the
summary statistics. This must be the result from a call to
\code{\link[=bootstrap_options]{bootstrap_options()}}.}
}
\value{
A list.
}
\description{
Options for summary statistics
}
