% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{vec_assert_or_null}
\alias{vec_assert_or_null}
\title{Assert an argument has known prototype and/or size or is NULL}
\usage{
vec_assert_or_null(
  x,
  ptype = NULL,
  size = NULL,
  arg = caller_arg(x),
  call = caller_env()
)
}
\arguments{
\item{x}{A vector argument to check.}

\item{ptype}{Prototype to compare against. If the prototype has a
class, its \code{\link[vctrs:vec_ptype]{vec_ptype()}} is compared to that of \code{x} with
\code{identical()}. Otherwise, its \code{\link[=typeof]{typeof()}} is compared to that of
\code{x} with \code{==}.}

\item{size}{A single integer size against which to compare.}

\item{arg}{Name of argument being checked. This is used in error
messages. The label of the expression passed as \code{x} is taken as
default.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
Either throws an error or returns \code{x}, invisibly.
}
\description{
Assert an argument has known prototype and/or size or is NULL
}
