% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vpc.R
\name{vpc_options}
\alias{vpc_options}
\title{Options for VPC simulations}
\usage{
vpc_options(..., n_samples = 250, seed = 123, skip = FALSE, vpc_only = FALSE)
}
\arguments{
\item{...}{These dots are reserved for future extensibility and must be empty.}

\item{n_samples}{Number of iterations to perform for VPC simulations.}

\item{seed}{Seed for random number generation.}

\item{skip}{Logical. Skip VPC simulations?}

\item{vpc_only}{Logical. Only return VPC simulations?}
}
\value{
A list.
}
\description{
Options for VPC simulations
}
