% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_grouping_column.R
\name{add_grouping_column}
\alias{add_grouping_column}
\title{Add grouping based on time bins}
\usage{
add_grouping_column(data, fun = group_by_time, label = "group", ...)
}
\arguments{
\item{data}{NONMEM-style data.frame, or path to CSV file with NONMEM data}

\item{fun}{grouping function to use. Can be custom function, or one of the
built-in grouping functions: group_by_time, group_by_dose. If a custom
function, input to the function needs to take at least the arguments \code{.data}
, \code{dictionary} and \code{...}, and return a data.frame with the same number of
rows as the input (and of course a new column providing the grouping).}

\item{label}{column name of grouper column in dataset}

\item{...}{passed onto the chosen grouper function, see the respective
grouping functions for more details. E.g. for \code{group_by_time}, need a \code{bins}
argument.}
}
\value{
data.frame
}
\description{
Add grouping based on time bins
}
