% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_shrinkage.R
\name{calculate_shrinkage}
\alias{calculate_shrinkage}
\title{Function to calculate eta-shrinkage}
\usage{
calculate_shrinkage(res_df, mod_obj)
}
\arguments{
\item{res_df}{data.frame or tibble with raw results (including parameters)
from a run with \code{mipdeval::run_eval()}.}

\item{mod_obj}{list object with model information}
}
\value{
tibble
}
\description{
Shrinkage for population PK model was first defined in this paper by
Savic and Karlsson: https://pmc.ncbi.nlm.nih.gov/articles/PMC2758126/.
It is a measure of how much information is available to update individual
estimates away from the population value. In principle, if there is no
information at all (i.e. in the case of population estimates only),
shrinkage is 100\%. In the case of fully informed Bayesian estimates
(unlikely to be achieved in practice), shrinkage is 0\%. In most practical
scenarios with limited sampling, shrinkage will be between 10-30\%. When
shrinkage is higher than 50\% or so, one could argue there is limited benefit
of sampling at all. So the shrinkage reported in this package can be used
to evaluate whether the sampling was efficient, and how shrinkage could be
reduced with additional sampling.
}
