% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stats.R
\name{calculate_stats}
\alias{calculate_stats}
\title{Calculate basic statistics, like RMSE, MPE, MAPE for forecasted data}
\usage{
calculate_stats(res, rounding = 3)
}
\arguments{
\item{res}{output object (\code{mipdeval_results}) from \code{run_eval()}, or
\code{data.frame} with raw results.}

\item{rounding}{number of decimals to round to.}
}
\value{
tibble
}
\description{
Calculate basic statistics, like RMSE, MPE, MAPE for forecasted data
}
