% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_required_covariates}
\alias{get_required_covariates}
\title{Get required covariates from a PKPDsim model object}
\usage{
get_required_covariates(model_object)
}
\description{
Get required covariates from a PKPDsim model object
}
