% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eval_core.R
\name{handle_sample_weighting}
\alias{handle_sample_weighting}
\title{Handle weighting of samples}
\usage{
handle_sample_weighting(obs_data, iterations, incremental, i)
}
\arguments{
\item{iterations}{numeric vector of groups}

\item{incremental}{should MAP Bayesian do incremental fits in the iterative
loop? I.e. in this case it would use the first iterations MAP Bayesian
estimates as input for the second iteration, and so forth. The uncertainty
around the MAP estimates would be used as the new \code{omega} matrix. This
approach has been called "model predictive control (MPC)"
(www.page-meeting.org/?abstract=9076) and may be more predictive than
"regular" MAP in some scenarios. Default is \code{FALSE}.}

\item{i}{index}
}
\value{
vector of weights (numeric)
}
\description{
This function is used to select the samples used in the fit (1 or 0),
but also to select their weight, if a sample weighting strategy is
selected.
}
