% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_input_data.R
\name{parse_input_data}
\alias{parse_input_data}
\title{Parse NONMEM-style input data, prepare for main eval loop}
\usage{
parse_input_data(data, covariates, ids, group = NULL, verbose = TRUE)
}
\arguments{
\item{data}{NONMEM-style data.frame, or path to CSV file with NONMEM data}

\item{covariates}{vector of covariate names}

\item{ids}{optional, vector of subject IDs to run analysis on (by default
runs analysis on all subjects in dataset)}

\item{group}{name of column in \code{data} that groups observations together in
iterative flow. By default each observation will be its own "group", but
this can be used to group peaks and troughs together, or to group
observations on the same day together. Grouping will be done prior to
running the analysis, so cannot be changed afterwards.}

\item{verbose}{show more output}
}
\value{
list of lists, within each list regimen, observations, and
covariates
}
\description{
Parse NONMEM-style input data, prepare for main eval loop
}
