% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_input_data.R
\name{parse_input_data}
\alias{parse_input_data}
\title{Parse NONMEM-style input data, prepare for main eval loop}
\usage{
parse_input_data(data, covariates)
}
\arguments{
\item{data}{NONMEM-style input datataset, as data.frame}

\item{covariates}{vector of covariate names}
}
\value{
list of lists, within each list regimen, observations, and
covariates
}
\description{
Parse NONMEM-style input data, prepare for main eval loop
}
