% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_model_info.R
\name{parse_model}
\alias{parse_model}
\alias{parse_model.character}
\alias{parse_model.PKPDsim}
\title{Parse PKPDsim model information}
\usage{
parse_model(model, ...)

\method{parse_model}{character}(
  model,
  ...,
  parameters = NULL,
  ruv = NULL,
  omega = NULL,
  iov_bins = NULL
)

\method{parse_model}{PKPDsim}(model, ..., parameters, ruv, omega, iov_bins)
}
\arguments{
\item{model}{either a PKPDsim model object, or a string pointing to a
PKPDsim-generated model library, e.g. \code{pkvancothomson}}

\item{...}{Arguments passed to methods.}

\item{parameters}{list of parameters}

\item{ruv}{residual error variability magnitude, specified as list.}

\item{omega}{between subject variability, supplied as vector specifiying the
lower triangle of the covariance matrix of random effects}

\item{iov_bins}{bins for inter-occasion variability. Passed unchanged to 
PKPDsim.}
}
\value{
A named list.
}
\description{
Parse model information from PKPDsim model object or an installed PKPDsim
model library.
}
