% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_psn_proseval_results.R
\name{parse_psn_proseval_results}
\alias{parse_psn_proseval_results}
\title{Parse PsN::proseval results.csv to filter out only the rows
that we need (for prediction of next level)}
\usage{
parse_psn_proseval_results(data)
}
\arguments{
\item{data}{results data.frame or path to csv file}
}
\description{
Parse PsN::proseval results.csv to filter out only the rows
that we need (for prediction of next level)
}
