% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eval.R
\name{run_eval}
\alias{run_eval}
\title{Run iterative predictive analysis, looping over each individual's data}
\usage{
run_eval(
  model,
  data,
  parameters = NULL,
  omega = NULL,
  error = NULL,
  groups = NULL,
  weights = NULL,
  prior_weight = 1,
  leak_covariates = FALSE,
  incremental = FALSE,
  threads = 1
)
}
\arguments{
\item{model}{model, created using `PKPDsim::new_ode_model()`}

\item{data}{NONMEM-style data.frame, or path to CSV file with NONMEM data}

\item{parameters}{list of parameters}

\item{omega}{between subject variability, supplied as vector specifiying the
lower triangle of the covariance matrix of random effects}

\item{error}{residual error, specified as list with arguments `add` and/or 
`prop` specifying the additive and proportional parts}

\item{groups}{variable in dataset that groups observations together in
iterative flow. By default each observation will be its own "group", but
this can be used to group peaks and troughs together, or to group
observations on the same day together.}

\item{weights}{vector of weights for error. Length of vector should be same 
as length of observation vector. If NULL (default), all weights are equal. 
Used in both MAP and NP methods. Note that `weights` argument will also 
affect residuals (residuals will be scaled too).}

\item{leak_covariates}{with the \code{proseval} tool in PsN, there is “data
leakage” (of the future covariates): since the NONMEM dataset in each step
contains the covariates for the future, this technically data leakage,
and could result in a too optimistic estimate of predictive performance.
In \code{mipdeval}, this is switched off by default (no data leakage), but it
can be switched on if we want to match the behavior of \code{PsN::proseval}
exactly.}

\item{incremental}{should MAP Bayesian do incremental fits in the iterative
loop. I.e. in this case it would use the first iterations MAP Bayesian
estimates as input for the second iteration, and so forth. The uncertainty
around the MAP estimates would be used as the new \code{omega} matrix. This
approach has been called "model predictive control (MPC)"
(www.page-meeting.org/?abstract=9076) and may be more predictive than
"regular" MAP in some scenarios. Default is \code{FALSE}.}

\item{threads}{number of threads to divide computations on. Default is 1,
i.e. no parallel execution}
}
\description{
Run iterative predictive analysis, looping over each individual's data
}
