% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eval.R
\name{run_eval}
\alias{run_eval}
\title{Run iterative predictive analysis, looping over each individual's data}
\usage{
run_eval(
  model,
  data,
  parameters = NULL,
  omega = NULL,
  iov_bins = NULL,
  ruv = NULL,
  dictionary = list(),
  groups = NULL,
  weights = NULL,
  weight_prior = 1,
  censor_covariates = TRUE,
  incremental = FALSE,
  threads = 1,
  progress = TRUE
)
}
\arguments{
\item{model}{either a PKPDsim model object, or a string pointing to a
PKPDsim-generated model library, e.g. \code{pkvancothomson}}

\item{data}{NONMEM-style data.frame, or path to CSV file with NONMEM data}

\item{parameters}{list of parameters}

\item{omega}{between subject variability, supplied as vector specifiying the
lower triangle of the covariance matrix of random effects}

\item{iov_bins}{bins for inter-occasion variability. Passed unchanged to 
PKPDsim.}

\item{ruv}{residual error variability magnitude, specified as list.}

\item{dictionary}{data dictionary. Optional, a list that specifies the
column names to be used from the dataset.}

\item{groups}{variable in dataset that groups observations together in
iterative flow. By default each observation will be its own "group", but
this can be used to group peaks and troughs together, or to group
observations on the same day together. Grouping will be done prior to
running the analysis, so cannot be changed afterwards.}

\item{weights}{vector of weights for error. Length of vector should be same 
as length of observation vector. If NULL (default), all weights are equal. 
Used in both MAP and NP methods. Note that `weights` argument will also 
affect residuals (residuals will be scaled too).}

\item{weight_prior}{weighting of priors in relationship to observed data, 
default = 1}

\item{censor_covariates}{with the \code{proseval} tool in PsN, there is “data
leakage” (of future covariates data): since the NONMEM dataset in each step
contains the covariates for the future, this is technically data leakage,
and could result in an over-optimistic estimate of predictive performance.
In \code{mipdeval}, covariate censoring of future covariate data is switched on
by default (so no data leakage), but it can be switched off if we want to
match the behavior of \code{PsN::proseval} exactly.}

\item{incremental}{should MAP Bayesian do incremental fits in the iterative
loop. I.e. in this case it would use the first iterations MAP Bayesian
estimates as input for the second iteration, and so forth. The uncertainty
around the MAP estimates would be used as the new \code{omega} matrix. This
approach has been called "model predictive control (MPC)"
(www.page-meeting.org/?abstract=9076) and may be more predictive than
"regular" MAP in some scenarios. Default is \code{FALSE}.}

\item{threads}{number of threads to divide computations on. Default is 1,
i.e. no parallel execution}

\item{progress}{should a progress bar be shown? Default is \code{TRUE}, but when
debugging the package it is useful to have it off, since progress bar
handlers obscure R output.}
}
\description{
Run iterative predictive analysis, looping over each individual's data
}
