% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vpc.R
\name{run_vpc_core}
\alias{run_vpc_core}
\title{Core function for \code{run_vpc}. Runs \code{n_samples} simulations for
a single subject}
\usage{
run_vpc_core(data, mod_obj, progress_function, n_samples = 250)
}
\arguments{
\item{data}{NONMEM-style data.frame, or path to CSV file with NONMEM data}

\item{mod_obj}{list object with model information}

\item{progress_function}{function to increment progress bar}

\item{n_samples}{number of iterations to perform for VPC simulations.}
}
\value{
data.frame
}
\description{
Core function for \code{run_vpc}. Runs \code{n_samples} simulations for
a single subject
}
